Change sample_rate to use u32

+++ crates/dawremi_core/src/effects/autotune.rs
@@ -12,7 +12,7 @@ pub enum AutotuneConfig {
 }
 
 pub struct Autotune {
-    pub sample_rate: f64,
+    pub sample_rate: u32,
     // TODO This can be changed for a list of numbers, to split in different sizes
     pub beat_length: usize,
     pub config: AutotuneConfig,
